refactor: devide some components
