+++ landing/fragments/landing-navigation/src/drawer/index.ts
@@ -1,1 +1,2 @@
 export * from './drawer.component'
+export * from './drawer-mobile.component'
