+++ landing/fragments/landing-navigation/src/index.ts
@@ -1,1 +1,2 @@
 export * from './navigation.component'
+export * from './navigation-mobile.component'
