+++ landing/fragments/landing-navigation/src/item/index.ts
@@ -1,1 +1,2 @@
 export * from './item.component'
+export * from './item-drawer.component'
