+++ landing/fragments/landing-navigation/src/navigation.component.tsx
@@ -14,7 +14,6 @@ import { Text }             from '@ui/text'
 
 import { NavLinks }         from './data'
 import { DrawerDesktop }    from './drawer'
-import { DrawerMobile }     from './drawer'
 import { Item }             from './item'
 import { NavigationProps }  from './navigation.interfaces'
 
@@ -64,47 +63,6 @@ export const NavigationDesktop: FC<NavigationProps> = ({ sectionRefs }: Navigati
         </Box>
 
         <Layout flexBasis={40} />
-      </Row>
-    </>
-  )
-}
-
-export const NavigationMobile = () => {
-  const [active, setActive] = useState<boolean>(false)
-
-  const handleClick = () => setActive(!active)
-
-  return (
-    <>
-      <DrawerMobile active={active} onClose={handleClick} />
-
-      <Row height={80} alignItems='center' justifyContent='center'>
-        <Layout flexBasis={20} />
-
-        <Box>
-          <Logo fill='white' width={40} height={40} />
-        </Box>
-
-        <Layout flexBasis={195} flexGrow='1' flexShrink='10' />
-
-        <Box width={100}>
-          <Button
-            onClick={handleClick}
-            variant='ghostBackgroundWhiteText'
-            size='smallSizeLittleRadii'
-            icon={<ArrowDownIcon width={9} height={5} />}
-            widthIcon={28}
-            heightIcon={28}
-            backgroundIcon='background.white'
-            radiiIcon='atom'
-          >
-            <Text color='white' fontSize='tiny' fontWeight='normal' lineHeight='small'>
-              <FormattedMessage id='navigation.button' />
-            </Text>
-          </Button>
-        </Box>
-
-        <Layout flexBasis={20} />
       </Row>
     </>
   )
